Remove limit of transaction group count
